feat: Update ArtifactoryAnalyser to use apache HTTPClient5 like the reset of DependencyCheck
Also mitigate issue #5868 by skipping (with a warning log) unusable results instead of throwing an exception so that multi-result responses can be handled.
